Split usb driver
